requirements.txt
psycopg[binary]>=3.2
jinja2>=3.1.6
asfeslib>=0.3.1
pydantic_settings>=2.12
bcrypt>=5.0
PyJWT>=2.10.1